+++ tests/test_models.py
@@ -17,6 +17,7 @@ EVENT_INTELLIGENCE_FOUNDATION_TABLES = {
     "event_candidates",
     "event_candidate_evidence",
     "impact_analyses",
+    "event_fact_snapshots",
 }
 
 FUTURE_TABLE_NAMES = {